Refactor container ID validation into utils module
Fix security risk by moving validation logic to a shared utility module.
Removes duplicate implementations of validate_container_identifier from
api_proxy.py and app.py, ensuring consistent validation across the codebase.
